rootless: Change the group ownership and group rw permissions for image
The current user needs to have read+write permissions for the kata image
to be able to use this as an nvdimm device.
Instead of opening up the permissions for "all", change the group
ownership of the image to "kvm" and open up the permissions for just the
"kvm" group.
This is something we should consider doing as part of packaging and
kata-deploy in the future maybe.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>